Console.log missing from reducer in master
The README states, "There is a console.log in our reducer that displays actions. Clicking the delete button should log an action with the todo's text content as the payload."

This console.log is not present in the reducer, manageTodo.js, when the project is initially cloned down. Please add this console.log in to make the codealong easier to follow.

Best wishes,
Kayla